student-housing-platform/requirements.txt
Flask
flask-mysqldb
razorpay
PyPDF2
python-dotenv